requirements.txt
alembic==1.6.0
appdirs==1.4.4
black==21.4b2
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
fastapi==0.63.0
filelock==3.0.12
greenlet==1.0.0
h11==0.12.0
httptools==0.1.2
identify==2.2.4
isort==5.8.0
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.8.1
pre-commit==2.12.1
psycopg2==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.4.12
starlette==0.13.6
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
virtualenv==20.4.4
watchgod==0.7
websockets==8.1
